Move stack_as_rebase_steps function
